- fixed: FBufferedUniform1f didn'T work because it used an int as its buffered value.

+++ src/gl/shaders/gl_shader.h
@@ -80,7 +80,7 @@ public:
 
 class FBufferedUniform1f
 {
-	int mBuffer;
+	float mBuffer;
 	int mIndex;
 
 public:
